Fix BaseResourceModel.save to not assume that options is non-null.
In the vast majority of cases, calling `save()` on a resource will
either include a caller-provided `options` object, or will get the
default value from being unspecified. The one case I've found where this
is not the case is `Backbone.Collection.create()`, which explicitly
calls `save()` on the object with a `null` options.

This change bullet-proofs that so we don't crash.

Testing Done:
Saw that I could use code that called `Collection.create` without it
hitting an error.

Reviewed at https://reviews.reviewboard.org/r/13612/