Add image resolution menu to newly-added images in the diff viewer.
The image review UI has something of a UI quirk about where we stick the
resolution drop-down menu. For attachments, we either stick it in the
diff mode bar (when viewing a diff), or the caption (when viewing a
single revision). This didn't work when the review UI was embedded
inline in the diff viewer. Actual diffs for images worked fine, because
we had the diff mode bar, but newly-added image files didn't (since
there's no mode bar and no separate caption bar). This change adds an
empty mode bar to use when we're in that situation.

Testing Done:
Viewed a diff that contained a newly-added image file. Saw that I had a
mostly-empty toolbar that allowed me to select the zoom for the image.

Reviewed at https://reviews.reviewboard.org/r/13628/