Comment serialization cleanup part 1: Break out basic definition.
The way comments are serialized for use in Review UIs and the diff
viewer is extremely messy. This change is the first step of cleaning
that up.

When I first converted the comment models to TypeScript, I mistakenly
marked the comment editor and comment dialog view's `publishedComments`
as being `BaseComment[]`. These are in fact the serialized comment data
arrays. This change moves the definition of the serialized comment
interface out of `abstractCommentBlockModel.ts` into its own file, where
we can import it anywhere we need, and makes use of that definition
where appropriate.

Testing Done:
Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/13655/