Add text/markdown as a supported mimetype for the Markdown UI.
More and more I'm seeing "text/markdown" as a MIME type for Markdown
files instead of the old "text/x-markdown". This change adds that as a
supported type for the review UI. This is a little bit moot because we
have some manual processing that assigns any `.md` file to
`text/x-markdown`, but we might as well be comprehensive.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13654/