Comment serialization cleanup part 4: Convert DiffFile to TypeScript.
The DiffFile model is quite simple, just being a data storage mechanism
for information about a file in a diff. This file uses a ton of terrible
names for things, and much of the documentation was just flat-out wrong.
This change converts it to TypeScript and spina, and corrects a bunch of
errors in the documentation about the attributes.

This also converts the DiffFileCollection class, which is extremely
small.

Testing Done:
Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/13658/