Add typing structures for the diff viewer context.
The context that's assembled and used to render the diff viewer page
(both the HTML-based UI and the diff context API endpoint) has grown
organically since the very earliest days of Review Board, and is a
pretty big hairy mess. In order to bring at least a little bit of sanity
to it, this change adds typing structures so that type checkers can
verify that we're always assigning things correctly.

The only actual change to the content of the context is that we can no
longer end up with non-`bool` types for the `is_draft_diff` and
`is_draft_interdiff` fields.

Testing Done:
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13666/