Update UI for the comments hint.
This change ports the comments hint model and view to TypeScript, and
updates the UI to include the new commit range information. Along with
this, the implementation has been simplified a bit to attach event
handlers individually to each item in the list, rather than having a
generic event handler and setting data on the list items.

Testing Done:
- Ran js-tests.
- Viewed a diff while having a review that had comments on lots of
  different revisions/interdiffs/commits. Saw that the comment hint
  correctly showed all of the revisions that had draft comments, and
  that clicking on all of them correctly loaded the associated revision.

Reviewed at https://reviews.reviewboard.org/r/13672/