Don't record changed file captions for diff file attachments.
Publishing updates to diffs that included file attachments was showing a
"File Captions" field in the change description that included captions
for every file in every commit in the new diff. These files are already
listed in the "Diff" field, and they don't have editable captions at all.

This change filters them out when we're deciding what to record in the
change description.

Testing Done:
Posted an update to a change that included binary files. Saw that there
was no longer a bunch of "File Caption" changes in the change description.

Reviewed at https://reviews.reviewboard.org/r/13673/