Fix computation of new-file state with commit ranges.
When selecting a range of commits where the first commit adds a file and
the later commit(s) change it, the `is_new_file` state was being
incorrectly computed as `True`. In the case where we have a base
filediff, we instead should be treating it as an existing file rather
than a new one.

Testing Done:
Loaded a commit range that was previously triggering the bug (especially
noticeable with the way binary files were loaded). Saw that I was now
seeing it treated as an existing file rather than new.

Reviewed at https://reviews.reviewboard.org/r/13674/