Select correct original file attachment when viewing commit ranges.
When viewing a range of commits, we weren't properly choosing the
original file attachment. This change makes it so if there's a
`base_filediff` present in the diff file information, we load the
attachment that is connected to that filediff.

Testing Done:
Posted a change that involved several commits to the same image file.
Selected various commit ranges and got the correct diffs for each commit
range.

Reviewed at https://reviews.reviewboard.org/r/13675/