Add file attachment comments to diff comments hint.
Now that we have file attachment comments showing up as part of the
diff, we also want to include that data in the diff comments hint. This
change adds that in, expanding the file attachment comment helper
function to return some additional data useful for this.

This simplifies the client side a bit by allowing the tip revision to be
null, for the case where we don't want the URL to include
`?tip-commit-id`.

Testing Done:
Created a draft review with a bunch of comments on files in various diff
revisions, interdiff revisions, and commit ranges (as well as some
comments on regular file attachments). Saw that the diff comment hint
correctly contained all the revisions and that clicking on everything
loaded the correct page.

Reviewed at https://reviews.reviewboard.org/r/13684/