Use Ink variables for link colors.
This replaces most of the usage of `@link-color` in the UI with
`var(--ink-p-link-color)` and `var(--ink-p-link-hover-color)`. These
give us consistent theme-friendly colors for links.

There are some instances that have not been replaced, but those will be
covered by larger changes.

Testing Done:
Tested most of the links. There's still a lot of work involved for these
bits of the UI.

Reviewed at https://reviews.reviewboard.org/r/13683/