Make ReviewUI generic for typing purposes.
This change adds generic typing for the ReviewUI class, allowing us to
specify in the subclass what the object type, comment type, and
serialized comment type are. This prevents the need for a bunch of
`cast` calls in the subclasses.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13696/