Consolidate font definitions and reduce font size overrides.
We now use Ink's font CSS variables and cease overriding or using
absolute font sizes wherever possible. This helps ensure consistency
throughout the UI.

Many of the font size overrides are there for legacy reasons, but aren't
really necessary anymore. In many of these cases we were just overriding
to what we expected the parent to use in order to override some style
that the selector picked up due to a precedence issue. We now just use
`inherit` where it's necessary, and remove the `font-size` where it's
not.

This doesn't mean that all font sizes are the same as what we had
before. The `9pt` font sizes are a bit bigger than our default. However,
what we have now is more consistent, and further refinements can be made
with relative font sizes or variables.

There are some font families and styles still set in places, but they're
part of larger cleanup changes, and are being ignored for now.

Testing Done:
Went through the affected parts of the UI and felt fine with the font
sizes. Made sure the font families were correct.

Reviewed at https://reviews.reviewboard.org/r/13688/