Tweak styles on forms in boxes to prevent some ugliness.
The OAuth application edit form uses our old `{% box %}` stuff, and a
conflict between the styling for boxes and the styling for `.rb-c-form`
was causing the background image in the box to shift up by the margin
added to the buttons. This change overrides the styling for the actions
to use padding instead of margin.

Testing Done:
Looked at the application edit form and saw that it wasn't ugly.