Set ReviewRequest.diffset_history to NULL on delete.
When a diffset history object is deleted, we don't actually want to
cascade and delete the review request. This change fixes that to set the
relation to NULL instead.

This is in preparation for a change that fixes it so we properly delete
the diffset history when deleting the review request (the opposite of
what the on_delete handler would previously do).

Testing Done:
Ran the evolution.

Reviewed at https://reviews.reviewboard.org/r/13734/