Add dark mode support for the review request box.
This introduces the beginnings of dark mode for the review request box.
In dark mode, the box takes on a cool-grey, almost blue-ish tone,
contrasting with the background but remaining dark and keeping the focus
on the content.

Most of the UI is converted over, but not all. The text boxes appear
light when in edit mode, and contrast quite a bit with the box, and the
file attachments and issue status area need new styling as well.
However, this gets us most of the way there.

There are some structural changes. The actions area is now styled using
a flexbox, instead of carefully managing floats. This helps us style
things just a bit more correctly, and will hopefully avoid some of the
off-by-one issues previously encountered on some platform/browser
combinations where the Reviews/Diff/Files tabs weren't fully connected
to the box.

There's also a fix to the field validation warning display. This appears
to have regressed at some point (likely Review Board 6 timeframe), where
the warning was populated with an attribute of the field verification
error, which didn't exist due to the error itself being a string. We'll
want to backport this later.

Future changes will introduce more styling to the contents of this box.

Testing Done:
Tested this on Chrome and Firefox in light, dark, and system color
schemes on both mobile and desktop views.

Tested field editing, validation, and action dropdowns to ensure the
styling was correct.

Reviewed at https://reviews.reviewboard.org/r/13729/